Small refactoring to CommentPageController
